Moved interface to component def
