+++ comp/anomalydetection/observer/impl/events.go
@@ -7,6 +7,7 @@ package observerimpl
 
 import (
 	observerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/def"
+	reporterdef "github.com/DataDog/datadog-agent/comp/anomalydetection/reporter/def"
 )
 
 // engineEventKind identifies the type of engine event.
@@ -63,14 +64,14 @@ type eventSink interface {
 // the event and active correlations from the stateView, then calls Report
 // on all registered reporters.
 type reporterEventSink struct {
-	reporters []observerdef.Reporter
+	reporters []reporterdef.Reporter
 	state     *stateView // for querying current correlations on advance
 }
 
 func (s *reporterEventSink) onEngineEvent(evt engineEvent) {
 	if evt.kind == eventAdvanceCompleted {
 		ac := evt.advanceCompleted
-		output := observerdef.ReportOutput{
+		output := reporterdef.ReportOutput{
 			AdvancedToSec: ac.advancedToSec,
 			NewAnomalies:  ac.anomalies,
 		}
