+++ comp/anomalydetection/reporter/def/component.go
@@ -7,7 +7,28 @@
 // detection events to the Datadog backend or stdout.
 package reporter
 
+import observerdef "github.com/DataDog/datadog-agent/comp/anomalydetection/observer/def"
+
 // team: q-branch
 
 // Component is the reporter component type.
 type Component interface{}
+
+// ReportOutput is the output model passed to reporters after each advance cycle.
+// It carries enough data for reporters to act without reaching back into engine internals.
+type ReportOutput struct {
+	// AdvancedToSec is the data time the engine advanced to.
+	AdvancedToSec int64
+	// NewAnomalies are anomalies detected in this advance cycle.
+	NewAnomalies []observerdef.Anomaly
+	// ActiveCorrelations are the current correlation patterns across all correlators.
+	ActiveCorrelations []observerdef.ActiveCorrelation
+}
+
+// Reporter receives reports and displays or delivers them.
+type Reporter interface {
+	// Name returns the reporter name for debugging.
+	Name() string
+	// Report delivers a report to its destination (stdout, file, webserver, etc).
+	Report(report ReportOutput)
+}
